requirements.txt
gym==0.19.0
keras==2.7.0
numpy==1.20.0
numba==0.54.1
pandas==1.3.4
scikit-learn==1.0.1
scipy==1.7.2
stable-baselines3==1.3.0
tensorflow-gpu==2.7.0
tensorflow-probability==0.12.2
torch==1.10.0
tqdm==4.62.3